Remove CultureService from the DI container
Since Avalonia seemingly doesn't support live-reload of static
strings, culture changes would only apply to dynamic content
created after the window loads. This would exclude constants, like
the menu bar, and would also affect anything already being displayed
when the culture changes.

Thusly, we are moving the culture service out of DI and making it
a one-off operation, at the launch of the program (before AXAML
files are parsed). This has the unfortunate side effect of not
allowing for live language switching, but it ensures the entire
application will use the selected language, so it's a worthy
trade-off.

If a future version of Avalonia updates how culture changes are
handled, we can re-visit this issue and bring the service back
to the DI container (with the Configuration update hook).